eqs: Swap thermal config for eqs-game-perf
The default thermal-engine-rtwo.conf is way too agressive with throttling that the performance is way too low for any users trying to play 3D games on a Snapdragon 8gen1 runs almost as slow as
a SD 888, so let's swap to the eqs-game-perf config as when we compared the configs it
allows for a better performance to thermal balance. We also checked eqs-perf but that one
was more of a let's get the highest scores in a benchmark and turn the device into a
handwarmer kind of thermal config.